Add logic to GameSystem to move blocks down. Add sample block to test scene.

+++ LD35GAME/src/Systems/TextureRenderSystem.cpp
@@ -9,8 +9,8 @@ void TextureRenderSystem::update(ex::EntityManager & em,
 
 	em.each<Body, GameBody>(
 		[this](ex::Entity entity, Body &body, GameBody &gameBody) {
-		body.position.x = gameBody.column * PIXELS_PER_GRID;
-		body.position.y = gameBody.row * PIXELS_PER_GRID;
+		body.position.x = static_cast<float> (gameBody.column * PIXELS_PER_GRID);
+		body.position.y = static_cast<float> (gameBody.row * PIXELS_PER_GRID);
 	});
 
 	em.each<Body, Sprite>(
